# src/gittest/swagger.yaml
openapi: 3.0.0
info:
  title: Marketplace API
  version: 1.0.0
servers: [
    {
      "url": "http://localhost:5000"
    }
  ]
paths:
...
  /api/v2/products:
    get:
      tags:
        - products
      summary: Get all products
      security:
        - jwt: []
      responses:
        '200':
          description: List of found products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Products not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
    post:
      tags:
        - products
      summary: Create a new product
      security:
        - jwt: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                seller_id:
                  type: integer
                title:
                  type: string
                price:
                  type: number
                rate:
                  type: number
              required:
                - seller_id
                - title
                - price
                - rate
      responses:
        '201':
          description: Product created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Product not created
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
  /api/v2/products/{product_id}:
    delete:
      tags:
        - products
      summary: Delete a product
      security:
        - jwt: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Product not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
    patch:
      tags:
        - products
      summary: Update a product
      security:
        - jwt: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                price:
                  type: number
                rate:
                  type: number
                seller_id:
                  type: integer
      responses:
        '200':
          description: Product updated
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Product not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
  /api/v2/products/id/{product_id}:
    get:
      tags:
        - products
      summary: Get product by ID
      security:
        - jwt: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
        '404':
          description: Product not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
  /api/v2/products/title/{title}:
    get:
      tags:
        - products
      summary: Get product by title
      security:
        - jwt: []
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
        '404':
          description: Product not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
...
  /api/v2/carts/{cart_id}/products:
    parameters:
      - name: cart_id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the cart
    get:
      tags:
        - products carts
      summary: Get products in cart by cart_id
      security:
        - jwt: []
      responses:
        '200':
          description: List of products in cart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductCart'
        '401':
          description: Unauthorized
        '404':
          description: Products in cart not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
  /api/v2/carts/{cart_id}/products/{product_id}:
    parameters:
      - name: cart_id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the cart
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the product
    post:
      tags:
        - products carts
      summary: Add product to cart
      security:
        - jwt: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
      responses:
        '201':
          description: Product added to cart
        '401':
          description: Unauthorized
        '404':
          description: Product not added to cart
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
    delete:
      tags:
        - products carts
      summary: Delete product from cart
      security:
        - jwt: []
      responses:
        '200':
          description: Product deleted from cart
        '401':
          description: Unauthorized
        '404':
          description: Product not found in cart
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
    patch:
      tags:
        - products carts
      summary: Update product in cart
      security:
        - jwt: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
      responses:
        '200':
          description: Product updated
        '401':
          description: Unauthorized
        '404':
          description: Product not found in cart
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Product:
      type: object
      properties:
        product_id:
          type: integer
        seller_id:
          type: integer
        title:
          type: string
        price:
          type: number
        rate:
          type: number
      required:
        - product_id
        - seller_id
        - title
        - price
        - rate
...
    ProductCart:
      type: object
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
        cart_id:
          type: integer
      required:
        - product_id
        - quantity
        - cart_id
...